refactor api tokens and fix all data mgt imports
