+++ src/app/api/data-mgt/http/entity-impl/sellers.http-data-api.service.ts
@@ -7,8 +7,8 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Seller } from 'src/app/models/entities/Seller';
-import { EntityHttpDataApiService } from './entity.http-data-api.aservice';
-import { EntityDataApiIService } from '../entity-data-api.iservice';
+import { EntityHttpDataApiService } from '../entity.http-data-api.aservice';
+import { EntityDataApiIService } from '../../entity-data-api.iservice';
 
 @Injectable()
 export class SellersHttpDataApiService
