+++ src/app/api/data-mgt/local-memory/entity-impl/products.local-memory-data-api.service.ts
@@ -6,11 +6,11 @@
 import { Injectable } from '@angular/core';
 import { ProductFilters } from 'src/app/shared/product-filters-panel/product-filters-panel.component';
 import { Product } from 'src/app/models/entities/Product';
-import { EntityLocalMemoryDataApiService } from './entity.local-memory-data-api.aservice';
+import { EntityLocalMemoryDataApiService } from '../entity.local-memory-data-api.aservice';
 import { Observable, of } from 'rxjs';
 import { ProductFamily } from 'src/app/models/entities/ProductFamily';
 import { ProductType } from 'src/app/models/entities/ProductType';
-import { MOCK_PRODUCT_FAMILIES, MOCK_PRODUCT_TYPES } from './shared.local-memory-data-api.service';
+import { MOCK_PRODUCT_FAMILIES, MOCK_PRODUCT_TYPES } from '../shared.local-memory-data-api.service';
 
 export const MOCK_PRODUCTS: Partial<Product>[] = [
   {
