+++ src/app/api/store/local-memory/local-memory-store-api.service.ts
@@ -10,7 +10,7 @@ import { StoreApiIService } from '../store-api.iservice';
 import { ProductFamily } from 'src/app/models/entities/ProductFamily';
 import { ProductType } from 'src/app/models/entities/ProductType';
 import { Product } from 'src/app/models/entities/Product';
-import { MOCK_PRODUCTS } from 'src/app/api/data-mgt/local-memory/products.local-memory-data-api.service';
+import { MOCK_PRODUCTS } from 'src/app/api/data-mgt/local-memory/entity-impl/products.local-memory-data-api.service';
 import { ProductFilters } from 'src/app/shared/product-filters-panel/product-filters-panel.component';
 import { MOCK_PRODUCT_TYPES, MOCK_PRODUCT_FAMILIES } from 'src/app/api/data-mgt/local-memory/shared.local-memory-data-api.service';
