+++ src/app/management/routes/clients/client-manager.service.ts
@@ -5,7 +5,7 @@
 
 import { Inject, Injectable } from '@angular/core';
 import { Client } from 'src/app/models/entities/Client';
-import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-tokens';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { EntityDataApiIService } from 'src/app/api/data-mgt/entity-data-api.iservice';
 import { DataManagerService } from '../data-manager.aservice';
