+++ src/app/management/routes/sales/sell-manager.service.ts
@@ -7,7 +7,7 @@ import { Inject, Injectable } from '@angular/core';
 import { Sell } from 'src/app/models/entities/Sell';
 import { SellDetail } from 'src/app/models/entities/SellDetail';
 import { CompositeEntityDataApiIService } from 'src/app/api/data-mgt/composite-entity-data-api.iservice';
-import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-tokens';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { DataManagerService } from '../data-manager.aservice';
 
 @Injectable()
