+++ src/app/management/routes/sellers/form-dialog/seller-manager-form.service.ts
@@ -4,7 +4,7 @@
 // https://opensource.org/licenses/MIT
 
 import { Inject, Injectable } from '@angular/core';
-import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-tokens';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { EntityDataApiIService } from 'src/app/api/data-mgt/entity-data-api.iservice';
 import { Seller } from 'src/app/models/entities/Seller';
 import { DataManagerFormService } from '../../data-manager-form.aservice';
