+++ src/app/management/routes/users/form-dialog/user-manager-form.service.ts
@@ -8,7 +8,7 @@ import { Observable } from 'rxjs';
 import { DataManagerFormService } from '../../data-manager-form.aservice';
 import { Person } from 'src/app/models/entities/Person';
 import { User } from 'src/app/models/entities/User';
-import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-tokens';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { EntityDataApiIService } from 'src/app/api/data-mgt/entity-data-api.iservice';
 
 @Injectable()
