+++ src/app/shared/product-filters-panel/product-filters-panel.service.spec.ts
@@ -7,7 +7,7 @@ import { TestBed } from '@angular/core/testing';
 import { of } from 'rxjs';
 import { ProductFiltersPanelService } from './product-filters-panel.service';
 import { StoreApiIService } from 'src/app/api/store/store-api.iservice';
-import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-tokens';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 
 describe('ProductFiltersPanelService', () => {
   let service: Partial<ProductFiltersPanelService>;
